# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Comunidade Alunos",
    "description": "Documentação completa das APIs do sistema de alunos e comunidade, incluindo o sistema de perguntas dinâmicas com histórico de respostas.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor Principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/posts": {
      "get": { "tags": ["Comunidade"], "summary": "Lista posts", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": { "tags": ["Comunidade"], "summary": "Cria post", "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Criado" } } }
    }
  }
}